fix(recommendations): order half-numbered series entries by fractional position
A 200-book run recommended Expanse novellas "Drive (#2.7)", "Gods of Risk
(#2.5)", and "The Vital Abyss (#5.5)" out of series order, ahead of the
unread next book "Caliban's War (#2)".

Series-position parsing was integer-only: the title regex captured `\d+`
and stopped at the decimal point, and the metadata path's `int()` raised on
"2.7". Fractional novellas therefore parsed as non-series items, dodged the
"too far ahead" suppression, and floated up on raw similarity.

Capture an optional fractional part in the series patterns, parse positions
with `float()`, and migrate series tracking from `set[int]` to `set[float]`
throughout. Rework `should_recommend_item`'s gap logic from integer
`range()` scans to a set-based "earliest unconsumed position" rule so a
`#2.5` novella waits until its preceding integer book is read. A
`math.isfinite` guard rejects malformed "inf"/"nan" metadata positions that
`float()` would otherwise accept.

+++ src/recommendations/scorers.py
@@ -92,7 +92,7 @@ class ScoringContext:
 
     preferences: UserPreferences
     consumed_items: list[ContentItem]
-    series_tracking: dict[str, set[int]]
+    series_tracking: dict[str, set[float]]
     content_type: ContentType
     all_unconsumed_items: list[ContentItem]
